Add: utils folder
utils folder was added and the constants mixed in the util.js were separated into another file(constants.js)
