+++ src/js/utils/constants.js
@@ -1,0 +1,26 @@
+const WEBTOON_URL = 'http://localhost:3001/webtoon/';
+
+const CL = {
+  SELECTED: 'selected',
+  WHOLE: 'whole',
+  DAY_ITEM: '.day__item',
+};
+
+const TITLE = {
+  promotion: '🔥HOT 최근 프로모션 진행작',
+};
+
+const VAL = {
+  CLONENODE_COUNT: 2,
+  TRANSITION_TIME: 0.25,
+  SWICHING_TIME: 500,
+  bannerWidth: 720,
+  MAX_BANNER_COUNT: 7,
+  MIN_BANNER_COUNT: 2,
+};
+
+const today = new Date()
+  .toLocaleString('en-us', { weekday: 'short' })
+  .toLowerCase();
+
+export { WEBTOON_URL, CL, TITLE, VAL, today };
